# Covid 2030 Game - Term 1 Project
**Course**: Computer Engineering Basics
**Instructor**: Professor Davood Abadi
**University**: University of Science and Technology

This project was developed one year ago, but due to the inexperience at the time, Git was not utilized. The entire project has now been uploaded.

The game is a terminal-based game, and some compilers may not display the ASCII art of the project properly. In the future, I may work on developing a GUI version of the game.

### How to Play:
The instructions and how-to-play guide can be found in the attached PDF file.

Feel free to explore and run the game!